+++ lib/appfuel/application.rb
@@ -1,3 +1,4 @@
-require_relative 'application/root'
+require_relative 'application/dispatcher'
 require_relative 'application/app_container'
+require_relative 'application/root'
 require_relative 'application/container_class_registration'